Major Update
- Updated Data file to remove Frag Belt from all specs
- Added the Shadow Cleave to the profile! This re-awakened profile
should come in handy when you want to put some extra emphasis on dotting
adds!
- Re-worked Shadow Word: Death priorities for the Divine Insight and
Solace and Insanity talents. It should now properly determine when to
double Shadow Word: Pain, when to use the first charge, when to use the
second charge and when to skip the second change completely.
- Fixed various minor bugs that lead to DPS losses across the board with
Vampiric Touch, Shadow Word: Pain and Mind Flay.
- Spamming Shadow Word: Pain while running should now properly account
for a SWP affected by UVLS on your target.
- All spells and abilities are now affected by Solace and Insanity.
Slightly altering their usage, behaviour and priority depending on this.

Other PvE additions:
- Added in Huolon farming function! This will automatically target and
start attacking Huolon for you if you're in range.