Major update
Seeing the commends, bugs and reports of the previous profile, I took a
step back and figured out just what I wanted to get out of the profile
itself. It's now fully optimized for speed and maximum damage potential.
The responsiveness of the profile should be quite noticeable from the
previous version(s), even my old pre-MoP one!

I've cut back on every unneeded check and the profile is now only
checking when it really needs to.

- All abilities have been optimized for speed.
- Shadow Word: Death have been re-written to judge when to double-tap
instantly, use the first charge and when to skip the second charge!
- Mind Blast should no longer be stopping the profile from executing
properly.
- Vampiric Touch has been re-written to keep up a strong DoT throughout
it's uptime, refreshing as necessary. It will also make a judgement call
when to let it run out or quickly refresh it between Mind Blast and
Devouring Plague if the DoT power is low!
- Shadow Word: Pain has been re-written to keep up a strong DoT
throughout it's uptime. It's also more responsive of target swaps. It
should follow the same guidelines as Vampiric Touch! This can be
double-checked by you through AffDots: Shadow Priest =)
- Mind Flay and Mind Flay: Insanity have recieved a few new changes. The
profile should now properly and ALWAYS break it's channel whenever
approperiate. After a tick, or right after a tick. This can be checked
by you through Gnosis!
- Mind Spike has been re-written to properly follow the spell DPET
hierarchy. It should also properly track when it has 2 stacks now and
never cap.
- Huolon Farming mode have been added! Park your ass at the bridge in
Timeless Isle, put the ability under [Misc] Profile init and AFK as you
please. (Remember to leave the profile running!)
- A Proving Grounds profile have been added for those of you who've
become as attached to PQR as I have. (Optimized for Gold and Endless
mode!)